Moved functional generalized test builder into submodule
