+++ settings.gradle
@@ -1,3 +1,4 @@
 rootProject.name = 'muisti-api'
 include 'api-spec'
 include 'api-client'
+include 'test-builder'
