+++ test-builder/build.gradle
@@ -1,0 +1,19 @@
+buildscript {
+    repositories {
+        jcenter()
+    }
+}
+
+apply plugin: 'kotlin'
+
+repositories {
+    jcenter()
+}
+
+dependencies {
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation project(':api-client')
+    implementation "fi.metatavu.jaxrs.testbuilder:jaxrs-functional-test-builder:$jaxrs_functional_test_builder_version"
+    implementation "org.awaitility:awaitility:$awaitility_version"
+    implementation 'commons-codec:commons-codec:1.13'
+}
