+++ test-builder/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'muisti-api-test-builder'
+include 'api-client'
