+++ test-builder/src/main/kotlin/fi/metatavu/muisti/api/test/builder/impl/DeviceModelTestBuilderResource.kt
